# openapi.yaml
openapi: 3.0.3
info:
  title: AFARI Platform API
  version: 0.1.0
  description: |
    The AFARI REST API. This document is the contract for the Phase 1
    foundation — the shape is stable and already implemented by the mock
    store at `src/lib/mock/store.ts`. Implementations for Phase 2+ land
    behind the same paths.
  contact:
    name: AFARI Platform
    url: https://afari.travel
servers:
  - url: https://api.afari.travel/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /trips/ai-parse:
    post:
      tags: [Trips]
      summary: Parse a natural-language trip description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: Lagos to Nairobi next Tuesday for a client meeting
      responses:
        "200":
          description: Structured trip draft
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  parsed:
                    type: object
                    properties:
                      origin: { type: object, nullable: true }
                      destination: { type: object, nullable: true }
                      departureDate: { type: string, format: date, nullable: true }
                      purpose: { type: string, nullable: true }
                  source: { type: string, enum: [heuristic, claude] }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...